Run eviction policy synchronously
I have no strong reason for the fire-and-forget approach. I don't really think it will cause any significant delays, and I'm not sure if it will be possible for it to run concurrently with itself in case there's more than one page loading. Probably not, but since it was not prepared for such scenario, I think it's better if it just runs asynchronously. Moreover, the code is easier to read this way.