Download scripts in parallel
Before this change, Witchcraft would synchronously go through each possible script, trying to download it and then injecting it.

The intention was to keep the logic simple. Since it has a predetermined script execution order, the easiest way is to just iterate them synchronously.

This change proposes a different approach where all scripts are downloaded in parallel, and then Witchcraft waits on a barrier (Promise.all) until all of them are finished. After the barrier is resolved, Witchcraft then proceeds to injecting them in order.

This is still relatively simple, but it still suffers from the problem that injection only starts after all scripts finish downloading.

This problem manifests more clearly if one of the scripts includes a remote file. In that case, scripts that have already finished downloading will have to wait until the remote one is complete so they can be injected. This would potentially make scripts miss the timing to run (e.g., the script expects to run at the very moment the document begins to be formed). While this is possibly not a real issue (users willing to run the script asap would avoid downloading remote code), fixing it is not that hard.

The fix involves downloading all the scripts in parallel, but then enqueueing their promises interleaved with their injection tasks such that the download promise comes first, and then the injection. So we'd have a queue similar to this:

    [download1, inject1, download2, inject2, ...]

The queue would be consumed in FIFO fashion, waiting for `download1` to complete before running `inject1`. Meanwhile, `download2` would already be running, but it would not be consumed from the queue until `inject1` would complete. That would guarantee that `inject2` would only run after `inject1` and `download2`.